requirements.txt
easyocr==1.7.1
imageio==2.31.3
numpy==1.24.3
opencv_contrib_python==4.8.0.76
opencv_python==4.8.0.76
opencv_python_headless==4.8.1.78
Pillow==9.5.0
Pillow==10.2.0
plotly==5.17.0
streamlit==1.28.0
streamlit_cropperjs==0.0.8
streamlit_image_select==0.6.0
streamlit_option_menu==0.3.6
tensorflow==2.13.0
tensorflow_intel==2.13.0
torch==2.0.1
torchvision==0.15.2
streamlit==1.24.0
streamlit==1.26.0
Requests==2.31.0
